Move enum to separate file.
